+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `StableVec::insert_into_hole()`
 - `StableVec::grow()`
 - `StableVec::clear()`
+- `Debug` implementations for `Iter`, `IterMut` and `Keys`
 
 ### Changed
 - The `Drop` impl now uses the `mem::needs_drop()` optimization hint to avoid